ES6: transform ommon/modules/ui/accessibility-prefs into ES6
